Return description with the rest of the conversation data
As the description is not typically needed in the conversation list, it
could be potentially "large" and the conversation list is fetched
frequently only a hash of the description is returned in that case. The
exception is when the description is empty, in which case an empty value
is returned, instead of the value of "sha1('')".

If only a single conversation is fetched the actual description is
always returned.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>